src/main: add optional syslog support
Instead of having to redirect (colored) logs to stdout/stderr, this
patch adds support for logging directly to syslog with approximate
GLogLevel to syslog level mapping.

By default all LOG_NOTICE, g_message(), and higher log messages are
logged to LOG_LOCAL0 facility.  This should of course be configurable
but is not at this stage.

Signed-off-by: Joachim Wiberg <troglobit@gmail.com>